Convert to jest in testAuthenticateRefreshToken.js
